# generated/openapi/app/zz_generated.types.go
// Package opapp provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package opapp

// Message defines model for Message.
type Message struct {
	Author    string `json:"author"`
	Content   string `json:"content"`
	MessageId int    `json:"message_id"`
	SpaceId   int    `json:"space_id"`
}

// Space defines model for Space.
type Space struct {
	Name    string `json:"name"`
	Owner   string `json:"owner"`
	SpaceId int    `json:"space_id"`
}

// User defines model for User.
type User struct {
	Id       int    `json:"id"`
	Name     string `json:"name"`
	Password string `json:"password"`
}

// CreateSpaceJSONRequestBody defines body for CreateSpace for application/json ContentType.
type CreateSpaceJSONRequestBody = Space

// CreateMessageJSONRequestBody defines body for CreateMessage for application/json ContentType.
type CreateMessageJSONRequestBody = Message

// RegisterUserJSONRequestBody defines body for RegisterUser for application/json ContentType.
type RegisterUserJSONRequestBody = User

// GetUserJSONRequestBody defines body for GetUser for application/json ContentType.
type GetUserJSONRequestBody = User